+++ src/test/java/com/epam/reportportal/testng/TestLaunchFinishShutdownHook.java
@@ -1,6 +1,6 @@
 package com.epam.reportportal.testng;
 
-import com.epam.reportportal.service.statistics.StatisticsService;
+import com.epam.reportportal.service.LaunchImpl;
 import com.epam.reportportal.testng.integration.feature.shutdown.LaunchFinishShutdownHookRemoveTest;
 import com.epam.reportportal.testng.integration.feature.shutdown.LaunchFinishShutdownHookTest;
 import com.epam.reportportal.util.test.ProcessUtils;
@@ -29,7 +29,7 @@ public class TestLaunchFinishShutdownHook {
 		ServerSocket ss = SocketUtils.getServerSocketOnFreePort();
 		SocketUtils.ServerCallable serverCallable = new SocketUtils.ServerCallable(ss, Collections.emptyMap(), "files/socket_response.txt");
 		Callable<Process> clientCallable = () -> ProcessUtils.buildProcess(true, clazz,
-				Collections.singletonMap(StatisticsService.DISABLE_PROPERTY, "1"), String.valueOf(ss.getLocalPort()));
+				Collections.singletonMap(LaunchImpl.DISABLE_PROPERTY, "1"), String.valueOf(ss.getLocalPort()));
 		Pair<List<String>, Process> startResult = SocketUtils.executeServerCallable(serverCallable, clientCallable, 15);
 		assertThat(startResult.getValue(), notNullValue());
 		assertThat("First request is a launch start", startResult.getKey().get(0), startsWith("POST /api/v2/test-project/launch"));